AI_Search_Knowledge_Graph/1_ai_search_loader/requirements.txt
pymssql==2.3.1
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-core==1.30.0
numpy>=1.24.0
